+++ src/components/common/neighbor/Button.jsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const Orange = styled.button`
+  width: 379px;
+  height: 40px;
+  border-radius: 4px;
+  outline: none;
+  color: white;
+  background-color: #ff8a3d;
+`;
+
+export const Span = styled.button`
+  font-size: 18px;
+`;
