+++ src/components/common/neighbor/Header.js
@@ -1,7 +1,4 @@
 import styled from 'styled-components';
-import { ReactComponent as SearchSvg } from '../../../icon/search.svg';
-import { ReactComponent as CloseSvg } from '../../../icon/close.svg';
-import ArrowBackImg from '../../../images/ico/ico_back_home_arrow.png';
 
 export const Wrapper = styled.div`
   display: flex;
@@ -9,13 +6,6 @@ export const Wrapper = styled.div`
   align-items: center;
   height: 92px;
   width: 100%;
-`;
-
-export const ArrowBackIcon = styled.img.attrs({
-  src: ArrowBackImg,
-})`
-  width: 19.33px;
-  cursor: pointer;
 `;
 
 export const SearchBarWrapper = styled.div`
@@ -27,27 +17,9 @@ export const SearchBarWrapper = styled.div`
   flex-flow: wrap;
 `;
 
-export const SearchIcon = styled(SearchSvg)`
-  width: 17px;
-  height: 17px;
-  fill: #98979c;
-`;
-
 export const Input = styled.input`
   width: 254px;
   height: 19px;
   border: none;
   outline: none;
-`;
-
-export const CloseIcon = styled(CloseSvg)`
-  width: 17px;
-  height: 17px;
-  cursor: pointer;
-`;
-
-export const Underline = styled.div`
-  width: 334px;
-  height: 2px;
-  background-color: #ebebed;
 `;
