+++ src/components/common/neighbor/Line.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Underline = styled.div`
+  width: 334px;
+  height: 2px;
+  background-color: #ebebed;
+`;
+
+export const Border = styled(Underline)`
+  width: 379px;
+`;
